requirements.txt
absl-py==1.4.0
astunparse==1.6.3
charset-normalizer==2.0.7
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django==5.0.4
dm-tree==0.1.8
etils==1.7.0
flatbuffers==24.3.25
fonttools==4.49.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
h5py==3.10.0
kagglehub==0.2.2
keras==3.5.0
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opt-einsum==3.3.0
optree==0.11.0
packaging==21.2
pandas==2.2.3
pillow==10.3.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.6
pyarrow==15.0.0
Pygments==2.17.2
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
readme-renderer==30.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
scipy==1.12.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-serving-api==2.17.0
tensorflow-text==2.16.1
termcolor==2.4.0
texttable==1.6.4
threadpoolctl==3.3.0
toml==0.10.2
twine==3.5.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==3.0.1
wrapt==1.14.1
